redo: rewrite redo apply (#2167)
ref pingcap/ticdc#1061

+++ tests/integration_tests/consistent_replicate_gbk/run.sh
@@ -17,21 +17,11 @@ export AWS_SECRET_ACCESS_KEY=$MINIO_SECRET_KEY
 export S3_ENDPOINT=127.0.0.1:24927
 rm -rf "$WORK_DIR"
 mkdir -p "$WORK_DIR"
-pkill -9 minio || true
-bin/minio server --address $S3_ENDPOINT "$WORK_DIR/s3" &
-MINIO_PID=$!
-i=0
-while ! curl -o /dev/null -v -s "http://$S3_ENDPOINT/"; do
-	i=$(($i + 1))
-	if [ $i -gt 30 ]; then
-		echo 'Failed to start minio'
-		exit 1
-	fi
-	sleep 2
-done
-
+MINIO_PID=0
 stop_minio() {
-	kill -2 $MINIO_PID || true
+	if [ $MINIO_PID -ne 0 ]; then
+		kill -2 $MINIO_PID || true
+	fi
 }
 
 stop() {
@@ -39,8 +29,6 @@ stop() {
 	stop_tidb_cluster
 	collect_logs $WORK_DIR
 }
-
-s3cmd --access_key=$MINIO_ACCESS_KEY --secret_key=$MINIO_SECRET_KEY --host=$S3_ENDPOINT --host-bucket=$S3_ENDPOINT --no-ssl mb s3://logbucket
 
 function run() {
 	# we only support eventually consistent replication with MySQL sink
@@ -50,8 +38,20 @@ function run() {
 
 	start_tidb_cluster --workdir $WORK_DIR
 
-	cd $WORK_DIR
+	bin/minio server --address $S3_ENDPOINT "$WORK_DIR/s3" &
+	MINIO_PID=$!
+	i=0
+	while ! curl -o /dev/null -v -s "http://$S3_ENDPOINT/"; do
+		i=$(($i + 1))
+		if [ $i -gt 30 ]; then
+			echo 'Failed to start minio'
+			exit 1
+		fi
+		sleep 2
+	done
+	s3cmd --access_key=$MINIO_ACCESS_KEY --secret_key=$MINIO_SECRET_KEY --host=$S3_ENDPOINT --host-bucket=$S3_ENDPOINT --no-ssl mb s3://logbucket
 
+	cd $WORK_DIR
 	run_cdc_server --workdir $WORK_DIR --binary $CDC_BINARY
 
 	SINK_URI="mysql://normal:123456@127.0.0.1:3306/"
@@ -85,10 +85,10 @@ function run() {
 	export GO_FAILPOINTS=''
 	export AWS_ACCESS_KEY_ID=$MINIO_ACCESS_KEY
 	export AWS_SECRET_ACCESS_KEY=$MINIO_SECRET_KEY
-	cdc redo apply --tmp-dir="$WORK_DIR/apply" \
+	cdc redo apply --log-level debug --tmp-dir="$WORK_DIR/apply" \
 		--storage="$storage_path" \
 		--sink-uri="mysql://normal:123456@127.0.0.1:3306/" >$WORK_DIR/cdc_redo.log
-	check_sync_diff $WORK_DIR $CUR/conf/diff_config.toml
+	check_sync_diff $WORK_DIR $CUR/conf/diff_config.toml 100
 }
 
 trap stop EXIT
